+++ dbmodel/schemas/main/common/base/fct/gw_fct_graphanalytics_mapzones_v1.sql
@@ -550,6 +550,11 @@ BEGIN
 					WHERE t.graphconfig IS NOT NULL
 					AND t.active
 					%s
+				), graphconfig_filtered AS (
+					SELECT g.* 
+					FROM graphconfig g
+					JOIN node n ON n.node_id = g.node_parent
+					WHERE EXISTS (SELECT 1 FROM vf_exploitation vfe WHERE vfe.expl_id = ANY(array_append(n.expl_visibility, n.expl_id)))
 				)
 				INSERT INTO temp_pgr_graphconfig (mapzone_id, graph_type, pgr_node_id, pgr_arc_id)
 				SELECT
@@ -557,7 +562,7 @@ BEGIN
 				'use',
 				node_parent,
 				to_arc
-				FROM graphconfig
+				FROM graphconfig_filtered
 				WHERE mapzone_id > 0
 				AND (node_parent IS DISTINCT FROM 0 OR to_arc IS DISTINCT FROM 0);
 			$sql$, v_mapzone_field, v_mapzone_table, v_query_text_aux);
@@ -681,6 +686,11 @@ BEGIN
 					JOIN LATERAL json_array_elements(t.graphconfig->'use') AS use_item ON TRUE
 					WHERE t.graphconfig IS NOT NULL
 					AND t.active
+				), graphconfig_filtered AS (
+					SELECT g.* 
+					FROM graphconfig g
+					JOIN node n ON n.node_id = g.node_parent
+					WHERE EXISTS (SELECT 1 FROM vf_exploitation vfe WHERE vfe.expl_id = ANY(array_append(n.expl_visibility, n.expl_id)))
 				)
 				INSERT INTO temp_pgr_graphconfig (mapzone_id, graph_type, pgr_node_id, pgr_arc_id)
 				SELECT
@@ -688,7 +698,7 @@ BEGIN
 					'use',
 					g.node_parent,
 					a.arc_id
-				FROM graphconfig g
+				FROM graphconfig_filtered g
 				LEFT JOIN v_temp_arc a ON a.node_2 = g.node_parent
 				WHERE g.mapzone_id > 0
 				AND g.node_parent IS DISTINCT FROM 0